tnsu ramp-up and updated ITE (#15)
Tensor-Network Simple-Update (tnsu, by Roi Elkabetz https://github.com/RoyElkabetz/Tensor-Networks-Simple-Update)  is used as a starting approximation for the PEPS near the ground state of the Kagome lattice under the Anti-Ferromagnetic (AFM) Heisenberg Model.

The Imaginary Time Evolution (ITE) Protocol, has also been improved for this task.



* option for constant edge order

* Update run_ite.py

* smaller health plot from log

* CACHE_CONTRACTION_ORDER

* working on creating a random valid density matrix

* always update best unit_cell

* ready for condor run

* dill->pickle and D>2 random state

* keep only single best unit_cell

* SAVE_FILES_WITH with if

* all_files in folder excludes subfolders

* revised random noise code

* condor ready

* pickle->dill

* _default_load_catch_other_load

* pickle->dill

* print logger name

* prepare for condor

* Update run_ite.py

* attempt to creare unit_cell from tnsu

* kagome in tnsu

* Update run_ite.py

* condor

* Update _tnsu.py

* new condor exec

* Update sender.py

* trying mean result from tnsu

* tnsu imports and newer python fix

* tnsu complex random tensors

* tnsu test

* tnsu absorb_all_weights

* Update tnsu_test.py

* Update sender.py

* longer tnsu test

* Update worker.py

* condor

* condor

* condor

* Update sender.py

* Update sender.py

* Update sender.py

* condor

* condor

* condor

* condor

* Update worker.py

* condor

* condor

* condor

* condor

* condor

* Update sender.py

* Update worker.py

* condor

* condor

* Update run_ite.py

* _working_on_tnsu

* working on tnsu

* tnsu ramp up

* condor

* condor

* Update sender.py

* Update worker.py

* Update sender.py

* condor

* Update run_ite.py

* Update run_ite.py

+++ src/libs/tnsu/structure_matrix_constructor.py
@@ -49,7 +49,7 @@ def infinite_structure_matrix_dict(name: str):
     return structure_matrix[name]
 
 
-def square_peps_pbc(side: np.int):
+def square_peps_pbc(side: int):
     """
     Creates a structure matrix of a square lattice Tensor network with periodic boundary conditions (pbc)
     of shape (side, side). The total number of tensors in the network would be side^2.
@@ -72,7 +72,7 @@ def square_peps_pbc(side: np.int):
     return structure_matrix
 
 
-def rectangular_peps_obc(height: np.int, width: np.int):
+def rectangular_peps_obc(height: int, width: int):
     """
     Creates a structure matrix of a rectangular lattice tensor network with open (non-periodic) boundary
     conditions (obc) of shape (height x width). The total number of tensors in the network would be height x width.
@@ -90,7 +90,7 @@ def rectangular_peps_obc(height: np.int, width: np.int):
                 edge_list.append((i, j, 4, i + 1, j, 2))
             if j < width - 1:
                 edge_list.append((i, j, 3, i, j + 1, 1))
-    structure_matrix = np.zeros(shape=[height * width, len(edge_list)], dtype=np.int)
+    structure_matrix = np.zeros(shape=[height * width, len(edge_list)], dtype=int)
 
     # fill in the structure matrix
     for edge_idx, edge in enumerate(edge_list):
